+++ flutter_appauth/ios/Classes/FlutterAppauthPlugin.h
@@ -4,5 +4,6 @@
 @interface FlutterAppauthPlugin : NSObject<FlutterPlugin>
 
 @property(nonatomic, strong, nullable) id<OIDExternalUserAgentSession> currentAuthorizationFlow;
+@property(nonatomic, strong, nullable) id<OIDExternalUserAgent> externalUserAgent;
 
 @end