+++ flutter_appauth_platform_interface/lib/src/authorization_request.dart
@@ -8,12 +8,12 @@ class AuthorizationRequest extends CommonRequestDetails
   AuthorizationRequest(
     String clientId,
     String redirectUrl, {
-    String? loginHint,
-    List<String>? scopes,
-    AuthorizationServiceConfiguration? serviceConfiguration,
-    Map<String, String>? additionalParameters,
     String? issuer,
     String? discoveryUrl,
+    AuthorizationServiceConfiguration? serviceConfiguration,
+    String? loginHint,
+    List<String>? scopes,
+    Map<String, String>? additionalParameters,
     List<String>? promptValues,
     bool allowInsecureConnections = false,
     bool preferEphemeralSession = false,
@@ -31,5 +31,6 @@ class AuthorizationRequest extends CommonRequestDetails
     this.allowInsecureConnections = allowInsecureConnections;
     this.preferEphemeralSession = preferEphemeralSession;
     this.responseMode = responseMode;
+    assertConfigurationInfo();
   }
 }